Begin integration with c-ares library
Without any new feature yet; this only sets the required scaffolding.

We plan to later use the c-ares capabilities to add an option to allow the
DNS server used for DNS resolution to be overridden, and an option to define
the list of default search domains. In both cases, the default (and actually
only possibility) right now is to use whatever is in /etc/resolv.conf

This is the first step in backporting Google-specific changes to upstream.

WARNING: the Windows-related change is untested, as I have no Windows system
to use for that. Some real testing would be appreciated.

Signed-off-by: Stefano Lattarini <slattarini@google.com>